@@ -1,43 +1,34 @@
 {
     "name": "becklyn/translations",
-    "type": "symfony-bundle",
     "description": "A bundle that extracts translations",
-    "homepage": "https://github.com/Becklyn/translations-bundle",
     "license": "BSD-3-Clause",
+    "type": "symfony-bundle",
     "authors": [
         {
             "name": "Becklyn Studios",
             "homepage": "https://becklyn.com/"
         }
     ],
+    "homepage": "https://github.com/Becklyn/translations-bundle",
     "require": {
         "php": ">=7.4",
         "ext-json": "*",
-        "becklyn/cache": "^1.0.1",
-        "becklyn/html-builder": "^2.1.1",
-        "becklyn/icon-loader": "^2.1.1",
-        "becklyn/rad": "^8.3.0",
-        "symfony/config": "^5.3",
-        "symfony/dependency-injection": "^5.3",
-        "symfony/framework-bundle": "^5.3",
-        "symfony/http-kernel": "^5.3",
-        "symfony/process": "^5.3",
-        "symfony/translation": "^5.3",
-        "symfony/translation-contracts": "^1.1 || ^2.0",
-        "twig/twig": "^2.11 || ^3.0"
+        "becklyn/cache": "^1.1.0",
+        "becklyn/html-builder": "^2.2.0",
+        "becklyn/icon-loader": "^2.2.0",
+        "becklyn/rad": "^8.6.0",
+        "symfony/config": "^5.4.3 || ^6.0.3",
+        "symfony/dependency-injection": "^5.4.3 || ^6.0.3",
+        "symfony/framework-bundle": "^5.4.3 || ^6.0.3",
+        "symfony/http-kernel": "^5.4.4 || ^6.0.4",
+        "symfony/process": "^5.4.3 || ^6.0.3",
+        "symfony/translation": "^5.4.3 || ^6.0.3",
+        "symfony/translation-contracts": "^1.1 || ^2.0 || ^3.0.0",
+        "twig/twig": "^3.3.8"
     },
     "require-dev": {
         "roave/security-advisories": "dev-master",
-        "symfony/phpunit-bridge": "^5.3"
-    },
-    "config": {
-        "sort-packages": true
-    },
-    "extra": {
-        "branch-alias": {
-            "dev-master": "1.x-dev",
-            "dev-next": "1.x-dev"
-        }
+        "symfony/phpunit-bridge": "^5.4.3 || ^6.0.3"
     },
     "autoload": {
         "psr-4": {
@@ -47,6 +38,15 @@
     "autoload-dev": {
         "psr-4": {
             "Tests\\Becklyn\\Translations\\": "tests/"
+        }
+    },
+    "config": {
+        "sort-packages": true
+    },
+    "extra": {
+        "branch-alias": {
+            "dev-master": "1.x-dev",
+            "dev-next": "1.x-dev"
         }
     }
 }
