@@ -8,6 +8,7 @@ use Becklyn\Translations\DependencyInjection\BecklynTranslationsConfiguration;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Extension\Extension;
+use Symfony\Component\DependencyInjection\Extension\ExtensionInterface;
 use Symfony\Component\DependencyInjection\Loader\YamlFileLoader;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
@@ -16,7 +17,7 @@ class BecklynTranslationsBundle extends Bundle
     /**
      * @inheritDoc
      */
-    public function getContainerExtension ()
+    public function getContainerExtension () : ?ExtensionInterface
     {
         return new class() extends Extension {
             /**
@@ -44,7 +45,7 @@ class BecklynTranslationsBundle extends Bundle
             /**
              * @inheritDoc
              */
-            public function getAlias ()
+            public function getAlias () : string
             {
                 return "becklyn_translations";
             }
